Implemented Tokens logic on seperate class and made some changes.
