+++ Contractor_Finding/API/Program.cs
@@ -24,6 +24,7 @@ option.UseSqlServer(connectionString)
 builder.Services.AddScoped<IUserService, UserService>();
 builder.Services.AddScoped<IEncrypt, Encrypt>();
 builder.Services.AddScoped<IContractorService, ContractorService>();
+builder.Services.AddScoped<IGenerateToken, GenerateToken>();
 builder.Services.AddScoped<ICustomerService, CustomerService>();
 
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
