+++ Contractor_Finding/Service/UserService.cs
@@ -21,11 +21,13 @@ namespace Service
         private readonly ContractorFindingContext contractorFindingContext;
         private readonly IEncrypt encrypt;
         private readonly IConfiguration _config;
-        public UserService(ContractorFindingContext contractorFindingContext, IEncrypt encrypt, IConfiguration config)
+        private readonly IGenerateToken generateToken;
+        public UserService(ContractorFindingContext contractorFindingContext, IEncrypt encrypt, IConfiguration config, IGenerateToken generateToken)
         {
             this.contractorFindingContext = contractorFindingContext;
             this.encrypt = encrypt;
             _config = config;
+            this.generateToken = generateToken;
         }
 
         //For Display
@@ -92,68 +94,9 @@ namespace Service
                 return "registration failed";
             }
         }
-        public ClaimsPrincipal? GetPrincipalFromExpiredToken(string? token)
-        {
-            var tokenValidationParameters = new TokenValidationParameters
-            {
-                ValidateAudience = true,
-                ValidateIssuer = true,
-                ValidateIssuerSigningKey = true,
-                ValidIssuer = _config["Jwt:Issuer"],
-                ValidAudience = _config["Jwt:Audience"],
-                IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_config["Jwt:Key"])),
-                ValidateLifetime = false
-            };
-
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var principal = tokenHandler.ValidateToken(token, tokenValidationParameters, out SecurityToken securityToken);
-            if (securityToken is not JwtSecurityToken jwtSecurityToken || !jwtSecurityToken.Header.Alg.Equals(SecurityAlgorithms.HmacSha256, StringComparison.InvariantCultureIgnoreCase))
-                throw new SecurityTokenException("Invalid token");
-
-            return principal;
-
-        }
-        public string? ValidateJwtToken(string token)
-        {
-            if (token == null)
-                return null;
-
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var key = Encoding.ASCII.GetBytes(_config["Jwt:Key"]);
-
-
-            try
-            {
-                tokenHandler.ValidateToken(token, new TokenValidationParameters
-                {
-                    ValidateIssuerSigningKey = true,
-                    IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_config["Jwt:Key"])),
-                    ValidateIssuer = true,
-                    ValidateAudience = true,
-                    ValidIssuer = _config["Jwt:Issuer"],
-                    ValidAudience = _config["Jwt:Audience"],
-
-                    // set clockskew to zero so tokens expire exactly at token expiration time (instead of 5 minutes later)
-                    ClockSkew = TimeSpan.Zero
-                }, out SecurityToken validatedToken);
-
-                var jwtToken = (JwtSecurityToken)validatedToken;
-                var userId = jwtToken.Claims.Select(x => x.Value).First();
-
-                // return user id from JWT token if validation successful
-                return userId;
-            }
-            catch (Exception ex)
-            {
-                // return null if validation fails
-                return null;
-            }
-        }
-
       
 
-
-
+      
         ////for Login
         public string Login(Login login)
         {
@@ -166,7 +109,7 @@ namespace Service
                 && u.Password == checkingpassword);
             if (myUser != null)
             {
-                var token = GenerateToken(myUser);
+                var token = generateToken.GenerateNewToken(myUser);
                 return token;
             }
             return null!;
@@ -174,28 +117,7 @@ namespace Service
 
         }
 
-        public string GenerateToken(TbUser user)
-        {
-            var securityKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_config["Jwt:Key"]!));
-            var credentials = new SigningCredentials(securityKey, SecurityAlgorithms.HmacSha256);
-            UserType role = contractorFindingContext.UserTypes.Where(x => x.TypeId == user.TypeUser).FirstOrDefault()!;
-            var claims = new[]
-            {
-                new Claim(ClaimTypes.NameIdentifier,user.EmailId!),
-                 new Claim(ClaimTypes.NameIdentifier,user.Password!),
-                  new Claim(ClaimTypes.Role, role.Usertype1!),
-
-            };
-            var token = new JwtSecurityToken(_config["Jwt:Issuer"],
-                _config["Jwt:Audience"],
-                claims,
-                expires: DateTime.Now.AddMinutes(15),
-                signingCredentials: credentials);
-
-
-            return new JwtSecurityTokenHandler().WriteToken(token);
-        }
-       
+     
 
         //for forgotpassword case
         public string forgotpassword(Login login)
